+++ Anima/filtering/denoising/non_local_means_image_filter/animaNLMeans.cxx
@@ -141,7 +141,7 @@ int main(int ac, const char** av)
         {
             std::cout<<"preparing filter..." << std::endl;
 
-            typedef itk::Image<float, 2> ImageType;
+            typedef itk::Image<double, 2> ImageType;
             typedef anima::NonLocalMeansImageFilter<ImageType> FilterType;
 
             FilterType::Pointer filter = FilterType::New();
@@ -180,7 +180,7 @@ int main(int ac, const char** av)
         {
             std::cout<<"preparing filter..." << std::endl;
 
-            typedef itk::Image<float, 3> ImageType;
+            typedef itk::Image<double, 3> ImageType;
             typedef anima::NonLocalMeansImageFilter<ImageType> FilterType;
 
             FilterType::Pointer filter = FilterType::New();
@@ -233,7 +233,7 @@ int main(int ac, const char** av)
 
             std::cout<<"preparing filter...\n";
 
-            typedef itk::Image<float, 4> ImageType;
+            typedef itk::Image<double, 4> ImageType;
             typedef anima::NonLocalMeansImageFilter<ImageType> FilterType;
 
             FilterType::Pointer filter = FilterType::New();
