+++ Anima/registration/resamplers/animaMCMResampleImageFilter.h
@@ -6,7 +6,7 @@
 namespace anima
 {
 
-template <typename TImageType, typename TInterpolatorPrecisionType=float>
+template <typename TImageType, typename TInterpolatorPrecisionType=double>
 class MCMResampleImageFilter :
         public OrientedModelBaseResampleImageFilter <TImageType, TInterpolatorPrecisionType>
 {
